fix eagle mediawiki bug
In commit `6d1ef15039ab6aea9588ec79936009de81de4e41`, the EAGLE
MediaWiki integration was broken. This is because the method
`CTSIdentifier.new_from_supplied` was changed but the code that called
it was not updated.

This commit updates the code in the `CtsPublicationsController` to use
the updated version of `new_from_supplied`.